Change mapping to gnomonic projection
The previous method for mapping RA/Dec to pixel coordinates was based on the
Mercator projection. Those resulting images are representative of the sky only
at low declination. At moderate declination (35 deg) -- even with a small FOV
(20 arcmin) -- the images are sufficiently distorted that they fail to be solved
by Astrometry.net.

This commit replaces that method with a gnomonic (tangent plane) projection,
which better approximates what a real camera pointed at the sky would see.